+++ src/app/controllers/serviceController.ts
@@ -1,5 +1,6 @@
 import { FastifyReply, FastifyRequest } from "fastify";
 import { z } from "zod";
+import { getUserId } from "../helper/getUserId";
 import { CompanyService } from "../services/companyService";
 import { ServiceService } from "../services/serviceService";
 
@@ -15,12 +16,18 @@ const schema = z.object({
 export class ServiceController {
   static async create(request: FastifyRequest, reply: FastifyReply) {
     const body = schema.parse(request.body);
+    const userId = getUserId(request);
+
     try {
       // Verify if company exists
       const company = await CompanyService.findById(body.companyId);
 
       if (!company) {
         return reply.code(404).send({ message: "Company not found" });
+      }
+
+      if (company.userId !== userId) {
+        return reply.code(403).send({ message: "Forbidden" });
       }
 
       const service = await ServiceService.create({
@@ -42,12 +49,18 @@ export class ServiceController {
       })
       .parse(request.query);
 
+    const userId = getUserId(request);
+
     try {
       // Verify if company exists
       const company = await CompanyService.findById(companyId);
 
       if (!company) {
         return reply.code(404).send({ message: "Company not found" });
+      }
+
+      if (company.userId !== userId) {
+        return reply.code(403).send({ message: "Forbidden" });
       }
 
       const services = await ServiceService.findAll(companyId);
@@ -64,12 +77,17 @@ export class ServiceController {
         id: z.string().uuid(),
       })
       .parse(request.params);
+    const userId = getUserId(request);
 
     try {
       const service = await ServiceService.findById(id);
 
       if (!service) {
         return reply.code(404).send({ message: "Service not found" });
+      }
+
+      if (service.company.userId !== userId) {
+        return reply.code(403).send({ message: "Forbidden" });
       }
 
       return reply.code(200).send(service);
@@ -84,8 +102,8 @@ export class ServiceController {
         id: z.string().uuid(),
       })
       .parse(request.params);
-
     const body = schema.parse(request.body);
+    const userId = getUserId(request);
 
     try {
       // Verify if company exists
@@ -99,6 +117,10 @@ export class ServiceController {
 
       if (!service) {
         return reply.code(404).send({ message: "Service not found" });
+      }
+
+      if (service.company.userId !== userId && company.userId !== userId) {
+        return reply.code(403).send({ message: "Forbidden" });
       }
 
       const updatedService = await ServiceService.update(id, {
@@ -119,11 +141,17 @@ export class ServiceController {
       })
       .parse(request.params);
 
+    const userId = getUserId(request);
+
     try {
       const service = await ServiceService.findById(id);
 
       if (!service) {
         return reply.code(404).send({ message: "Service not found" });
+      }
+
+      if (service.company.userId !== userId) {
+        return reply.code(403).send({ message: "Forbidden" });
       }
 
       await ServiceService.delete(id);
